prettier fix

+++ contracts/Market.sol
@@ -20,7 +20,7 @@ contract NFTMarket is ReentrancyGuard {
   }
 
   struct MarketItem {
-    uint itemId;
+    uint256 itemId;
     address nftContract;
     uint256 tokenId;
     address payable seller;
@@ -32,8 +32,8 @@ contract NFTMarket is ReentrancyGuard {
   mapping(uint256 => MarketItem) private idToMarketItem;
   mapping(uint256 => MarketItem) private tidToMarketItem;
 
-  event MarketItemCreated (
-    uint indexed itemId,
+  event MarketItemCreated(
+    uint256 indexed itemId,
     address indexed nftContract,
     uint256 indexed tokenId,
     address seller,
@@ -46,7 +46,7 @@ contract NFTMarket is ReentrancyGuard {
   function getListingPrice() public view returns (uint256) {
     return listingPrice;
   }
-  
+
   /* Places an item for sale on the marketplace */
   function createMarketItem(
     address nftContract,
@@ -59,7 +59,7 @@ contract NFTMarket is ReentrancyGuard {
     _itemIds.increment();
     uint256 itemId = _itemIds.current();
 
-    idToMarketItem[itemId] =  MarketItem(
+    idToMarketItem[itemId] = MarketItem(
       itemId,
       nftContract,
       tokenId,
@@ -69,7 +69,7 @@ contract NFTMarket is ReentrancyGuard {
       false
     );
 
-    tidToMarketItem[tokenId] =  MarketItem(
+    tidToMarketItem[tokenId] = MarketItem(
       itemId,
       nftContract,
       tokenId,
@@ -94,13 +94,17 @@ contract NFTMarket is ReentrancyGuard {
 
   /* Creates the sale of a marketplace item */
   /* Transfers ownership of the item, as well as funds between parties */
-  function createMarketSale(
-    address nftContract,
-    uint256 itemId
-    ) public payable nonReentrant {
-    uint price = idToMarketItem[itemId].price;
-    uint tokenId = idToMarketItem[itemId].tokenId;
-    require(msg.value == price, "Please submit the asking price in order to complete the purchase");
+  function createMarketSale(address nftContract, uint256 itemId)
+    public
+    payable
+    nonReentrant
+  {
+    uint256 price = idToMarketItem[itemId].price;
+    uint256 tokenId = idToMarketItem[itemId].tokenId;
+    require(
+      msg.value == price,
+      "Please submit the asking price in order to complete the purchase"
+    );
 
     idToMarketItem[itemId].seller.transfer(msg.value);
     IERC721(nftContract).transferFrom(address(this), msg.sender, tokenId);
@@ -114,18 +118,25 @@ contract NFTMarket is ReentrancyGuard {
     payable(owner).transfer(listingPrice);
   }
 
-  function createMarketSaleBytokenId(
-    address nftContract,
-    uint256 tokenId
-    ) public payable nonReentrant {
+  function createMarketSaleBytokenId(address nftContract, uint256 tokenId)
+    public
+    payable
+    nonReentrant
+  {
     // uint tokenId = tokenId; // TODO
     // Note: idToMarketItem maybe less than allItem(some item-minted not goto market)
 
-    uint itemId = tidToMarketItem[tokenId].itemId;
-    require(tokenId == idToMarketItem[itemId].tokenId, "Sorry! tokenId error, we will fix it");
-    uint price = tidToMarketItem[tokenId].price;
+    uint256 itemId = tidToMarketItem[tokenId].itemId;
+    require(
+      tokenId == idToMarketItem[itemId].tokenId,
+      "Sorry! tokenId error, we will fix it"
+    );
+    uint256 price = tidToMarketItem[tokenId].price;
 
-    require(msg.value == price, "Please submit the asking price in order to complete the purchase");
+    require(
+      msg.value == price,
+      "Please submit the asking price in order to complete the purchase"
+    );
 
     idToMarketItem[itemId].seller.transfer(msg.value);
     IERC721(nftContract).transferFrom(address(this), msg.sender, tokenId);
@@ -139,17 +150,16 @@ contract NFTMarket is ReentrancyGuard {
     payable(owner).transfer(listingPrice);
   }
 
-
   /* Returns all unsold market items */
   function fetchMarketItems() public view returns (MarketItem[] memory) {
-    uint itemCount = _itemIds.current();
-    uint unsoldItemCount = _itemIds.current() - _itemsSold.current();
-    uint currentIndex = 0;
+    uint256 itemCount = _itemIds.current();
+    uint256 unsoldItemCount = _itemIds.current() - _itemsSold.current();
+    uint256 currentIndex = 0;
 
     MarketItem[] memory items = new MarketItem[](unsoldItemCount);
-    for (uint i = 0; i < itemCount; i++) {
+    for (uint256 i = 0; i < itemCount; i++) {
       if (idToMarketItem[i + 1].owner == address(0)) {
-        uint currentId = i + 1;
+        uint256 currentId = i + 1;
         MarketItem storage currentItem = idToMarketItem[currentId];
         items[currentIndex] = currentItem;
         currentIndex += 1;
@@ -160,20 +170,20 @@ contract NFTMarket is ReentrancyGuard {
 
   /* Returns onlyq items that a user has purchased */
   function fetchMyNFTs() public view returns (MarketItem[] memory) {
-    uint totalItemCount = _itemIds.current();
-    uint itemCount = 0;
-    uint currentIndex = 0;
+    uint256 totalItemCount = _itemIds.current();
+    uint256 itemCount = 0;
+    uint256 currentIndex = 0;
 
-    for (uint i = 0; i < totalItemCount; i++) {
+    for (uint256 i = 0; i < totalItemCount; i++) {
       if (idToMarketItem[i + 1].owner == msg.sender) {
         itemCount += 1;
       }
     }
 
     MarketItem[] memory items = new MarketItem[](itemCount);
-    for (uint i = 0; i < totalItemCount; i++) {
+    for (uint256 i = 0; i < totalItemCount; i++) {
       if (idToMarketItem[i + 1].owner == msg.sender) {
-        uint currentId = i + 1;
+        uint256 currentId = i + 1;
         MarketItem storage currentItem = idToMarketItem[currentId];
         items[currentIndex] = currentItem;
         currentIndex += 1;
@@ -184,20 +194,20 @@ contract NFTMarket is ReentrancyGuard {
 
   /* Returns only items a user has created */
   function fetchItemsCreated() public view returns (MarketItem[] memory) {
-    uint totalItemCount = _itemIds.current();
-    uint itemCount = 0;
-    uint currentIndex = 0;
+    uint256 totalItemCount = _itemIds.current();
+    uint256 itemCount = 0;
+    uint256 currentIndex = 0;
 
-    for (uint i = 0; i < totalItemCount; i++) {
+    for (uint256 i = 0; i < totalItemCount; i++) {
       if (idToMarketItem[i + 1].seller == msg.sender) {
         itemCount += 1;
       }
     }
 
     MarketItem[] memory items = new MarketItem[](itemCount);
-    for (uint i = 0; i < totalItemCount; i++) {
+    for (uint256 i = 0; i < totalItemCount; i++) {
       if (idToMarketItem[i + 1].seller == msg.sender) {
-        uint currentId = i + 1;
+        uint256 currentId = i + 1;
         MarketItem storage currentItem = idToMarketItem[currentId];
         items[currentIndex] = currentItem;
         currentIndex += 1;
